# backend/donald/rest/openapi.yaml
...
openapi: 3.1.0
info:
    title: Donald
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
    /agent-edges:
        get:
            tags:
                - AgentEdge
            summary: List agent-edges
            operationId: listAgentEdge
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-edges
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdgeList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentEdge
            summary: Create AgentEdge
            operationId: createAgentEdge
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEdge'
                required: true
            responses:
                "201":
                    description: Created AgentEdge
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /agent-events:
        get:
            tags:
                - AgentEvent
            summary: List agent-events
            operationId: listAgentEvent
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-events
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEventList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentEvent
            summary: Create AgentEvent
            operationId: createAgentEvent
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEvent'
                required: true
            responses:
                "201":
                    description: Created AgentEvent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /clients:
        get:
            tags:
                - Client
            summary: List clients
            operationId: listClient
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of clients
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Client
            summary: Create Client
            operationId: createClient
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Client'
                required: true
            responses:
                "201":
                    description: Created Client
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Client'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /interventions:
        get:
            tags:
                - Intervention
            summary: List interventions
            operationId: listIntervention
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of interventions
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InterventionList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Intervention
            summary: Create Intervention
            operationId: createIntervention
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Intervention'
                required: true
            responses:
                "201":
                    description: Created Intervention
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Intervention'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
        AgentEdge:
            properties:
                condition_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                edge_type:
                    $ref: '#/components/schemas/AgentEdgeType'
                from_node_uuid:
                    format: uuid
                    type: string
                run_uuid:
                    format: uuid
                    type: string
                status:
                    $ref: '#/components/schemas/AgentEdgeStatus'
                to_node_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - from_node_uuid
                - to_node_uuid
                - edge_type
                - status
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
        AgentEdgeList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentEdge'
                    type: array
                next_page_token:
                    type: string
            type: object
        AgentEdgeStatus:
            enum:
                - invalid
                - pending
                - traversed
                - skipped
                - removed
            type: string
        AgentEdgeType:
            enum:
                - invalid
                - dependency
                - transition
                - conditional
                - retry
                - fallback
                - subagent_spawn
            type: string
        AgentEvent:
            properties:
                agent_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                event_type:
                    $ref: '#/components/schemas/AgentEventType'
                idempotency_key:
                    type: string
                node_uuid:
                    format: uuid
                    nullable: true
                    type: string
                occurred_at:
                    format: date-time
                    type: string
                payload:
                    $ref: '#/components/schemas/AgentEventPayload'
                    nullable: true
                run_uuid:
                    format: uuid
                    type: string
                sequence:
                    format: int64
                    type: integer
                status:
                    $ref: '#/components/schemas/AgentEventStatus'
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - sequence
                - event_type
                - idempotency_key
                - occurred_at
                - status
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
        AgentEventList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentEvent'
                    type: array
                next_page_token:
                    type: string
            type: object
        AgentEventStatus:
            enum:
                - invalid
                - recorded
                - discarded
            type: string
        AgentEventType:
            enum:
                - invalid
                - run_started
                - run_updated
                - run_finished
                - plan_declared
                - node_added
                - node_updated
                - node_status_changed
                - node_removed
                - edge_added
                - edge_updated
                - edge_removed
                - artifact_added
                - intervention_requested
                - intervention_delivered
                - intervention_resolved
                - heartbeat
                - agent_message
            type: string
...
        Client:
            properties:
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                industry:
                    $ref: '#/components/schemas/ClientIndustry'
                    nullable: true
                name:
                    type: string
                status:
                    $ref: '#/components/schemas/ClientStatus'
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - name
                - created_by
                - updated_by
                - status
                - created_at
                - updated_at
            type: object
        ClientIndustry:
            enum:
                - invalid
                - fintech
                - furniture
            type: string
        ClientList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Client'
                    type: array
                next_page_token:
                    type: string
            type: object
        ClientStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
        Intervention:
            properties:
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                delivered_at:
                    format: date-time
                    nullable: true
                    type: string
                expires_at:
                    format: date-time
                    nullable: true
                    type: string
                interactions:
                    items:
                        $ref: '#/components/schemas/InterventionInteraction'
                    nullable: true
                    type: array
                node_uuid:
                    format: uuid
                    type: string
                prompt:
                    nullable: true
                    type: string
                requested_by_uuid:
                    format: uuid
                    type: string
                resolved_at:
                    format: date-time
                    nullable: true
                    type: string
                resolved_by_uuid:
                    format: uuid
                    nullable: true
                    type: string
                run_uuid:
                    format: uuid
                    type: string
                status:
                    $ref: '#/components/schemas/InterventionStatus'
                status_message:
                    nullable: true
                    type: string
                type:
                    $ref: '#/components/schemas/InterventionType'
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - node_uuid
                - type
                - status
                - requested_by_uuid
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
...
        InterventionList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Intervention'
                    type: array
                next_page_token:
                    type: string
            type: object
        InterventionStatus:
            enum:
                - invalid
                - registered
                - failed
                - succeeded
                - picked_up_by_agent
                - expired
                - withdrawn
            type: string
        InterventionType:
            enum:
                - invalid
                - stop
                - steer
            type: string
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object