# backend/donald/rest/server/create_agent_edge.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	agent_edgemodule "github.com/nextwave/donald/core/module/agent_edge"
	"github.com/nextwave/donald/core/module/agent_edge/types"
	main_entity "github.com/nextwave/donald/entity/agent_edge"
)

func (h *handlers) CreateAgentEdge(w http.ResponseWriter, r *http.Request) {
	var item main_entity.AgentEdge
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.AgentEdge().Insert(r.Context(), types.UpsertRequest{
		AgentEdge: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.AgentEdge().FetchAgentEdgeByUUID(r.Context(), types.FetchAgentEdgeByUUIDRequest(res), agent_edgemodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.UUID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}