# backend/donald/rest/server/create_agent_event.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	agent_eventmodule "github.com/nextwave/donald/core/module/agent_event"
	"github.com/nextwave/donald/core/module/agent_event/types"
	main_entity "github.com/nextwave/donald/entity/agent_event"
)

func (h *handlers) CreateAgentEvent(w http.ResponseWriter, r *http.Request) {
	var item main_entity.AgentEvent
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.AgentEvent().Insert(r.Context(), types.UpsertRequest{
		AgentEvent: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.AgentEvent().FetchAgentEventByUUID(r.Context(), types.FetchAgentEventByUUIDRequest(res), agent_eventmodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.UUID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}