# backend/donald/rest/server/create_client.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	clientmodule "github.com/nextwave/donald/core/module/client"
	"github.com/nextwave/donald/core/module/client/types"
	main_entity "github.com/nextwave/donald/entity/client"
)

func (h *handlers) CreateClient(w http.ResponseWriter, r *http.Request) {
	var item main_entity.Client
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.Client().Insert(r.Context(), types.UpsertRequest{
		Client: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.Client().FetchClientByUUID(r.Context(), types.FetchClientByUUIDRequest(res), clientmodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.UUID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}