# backend/donald/rest/server/create_intervention.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	interventionmodule "github.com/nextwave/donald/core/module/intervention"
	"github.com/nextwave/donald/core/module/intervention/types"
	main_entity "github.com/nextwave/donald/entity/intervention"
)

func (h *handlers) CreateIntervention(w http.ResponseWriter, r *http.Request) {
	var item main_entity.Intervention
	if err := json.NewDecoder(r.Body).Decode(&item); err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	res, err := h.core.Intervention().Insert(r.Context(), types.UpsertRequest{
		Intervention: item,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	fetchRes, err := h.core.Intervention().FetchInterventionByUUID(r.Context(), types.FetchInterventionByUUIDRequest(res), interventionmodule.WithSkipCache())
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusInternalServerError, errors.New("entity not found after insert"))
		return
	}

	createdEntity := fetchRes.Results[0]

	// Location header
	location := fmt.Sprintf("%s/%v/", r.URL.Path, createdEntity.UUID)
	w.Header().Set("Location", location)

	writeJSON(w, http.StatusCreated, createdEntity)
}